+++ src/engine/engine_name.rs
@@ -1,5 +1,5 @@
 //! Port of: engine/engine_name.c
-//! IR hash: 699b5f0da57e8d78
+//! IR hash: 545f394232195ad9
 //! CODEGEN: signatures locked. Only fill todo!() bodies.
 
 use crate::types::*;
@@ -13,17 +13,17 @@ pub fn getnumadr(m: *const mjModel, r#type: mjtObj, padr: *mut *mut i32, mapadr:
     todo ! ()
 }
 
-/// C: mj_hashString (engine/engine_name.c:229)
+/// C: mj_hashString (engine/engine_name.h:30)
 #[allow(unused_variables, non_snake_case)]
-pub fn mj_hash_string(s: *const i8, n: i32) -> i32 {
+pub fn mj_hash_string(s: *const i8, n: u64) -> u64 {
     // WARNING: signature changed — verify body
-    // Previous params: (s : * const i8, n : i32)
-    // Previous return: i32
+    // Previous params: (s : * const i8, n : u64)
+    // Previous return: u64
     todo ! ()
 }
 
 /// C: mj_name2id (engine/engine_name.h:33)
-/// Calls: _getnumadr
+/// Calls: _getnumadr, mj_hashString
 #[allow(unused_variables, non_snake_case)]
 pub fn mj_name2id(m: *const mjModel, r#type: i32, name: *const i8) -> i32 {
     // WARNING: signature changed — verify body
